+++ src/Firebase/Exception/MessagingApiExceptionConverter.php
@@ -4,13 +4,17 @@ declare(strict_types=1);
 
 namespace Kreait\Firebase\Exception;
 
+use DateTimeImmutable;
 use GuzzleHttp\Exception\ConnectException;
 use GuzzleHttp\Exception\RequestException;
+use Kreait\Clock;
+use Kreait\Clock\SystemClock;
 use Kreait\Firebase\Exception\Messaging\ApiConnectionFailed;
 use Kreait\Firebase\Exception\Messaging\AuthenticationError;
 use Kreait\Firebase\Exception\Messaging\InvalidMessage;
 use Kreait\Firebase\Exception\Messaging\MessagingError;
 use Kreait\Firebase\Exception\Messaging\NotFound;
+use Kreait\Firebase\Exception\Messaging\QuotaExceeded;
 use Kreait\Firebase\Exception\Messaging\ServerError;
 use Kreait\Firebase\Exception\Messaging\ServerUnavailable;
 use Kreait\Firebase\Http\ErrorResponseParser;
@@ -25,12 +29,16 @@ class MessagingApiExceptionConverter
     /** @var ErrorResponseParser */
     private $responseParser;
 
+    /** @var Clock */
+    private $clock;
+
     /**
      * @internal
      */
-    public function __construct()
+    public function __construct(?Clock $clock = null)
     {
         $this->responseParser = new ErrorResponseParser();
+        $this->clock = $clock ?? new SystemClock();
     }
 
     /**
@@ -71,11 +79,20 @@ class MessagingApiExceptionConverter
             case 404:
                 $convertedError = new NotFound($message, $code, $previous);
                 break;
+            case 429:
+                $convertedError = new QuotaExceeded($message, $code, $previous);
+                if ($retryAfter = $this->getRetryAfter($response)) {
+                    $convertedError = $convertedError->withRetryAfter($retryAfter);
+                }
+                break;
             case 500:
                 $convertedError = new ServerError($message, $code, $previous);
                 break;
             case 503:
                 $convertedError = new ServerUnavailable($message, $code, $previous);
+                if ($retryAfter = $this->getRetryAfter($response)) {
+                    $convertedError = $convertedError->withRetryAfter($retryAfter);
+                }
                 break;
             default:
                 $convertedError = new MessagingError($message, $code, $previous);
@@ -92,5 +109,28 @@ class MessagingApiExceptionConverter
         }
 
         return new MessagingError($e->getMessage(), $e->getCode(), $e);
+    }
+
+    private function getRetryAfter(ResponseInterface $response): ?DateTimeImmutable
+    {
+        $retryAfter = $response->getHeader('Retry-After')[0] ?? null;
+
+        if (!$retryAfter) {
+            return null;
+        }
+
+        if (\is_numeric($retryAfter)) {
+            return $this->clock->now()->modify("+{$retryAfter} seconds");
+        }
+
+        try {
+            return new DateTimeImmutable($retryAfter);
+        } catch (Throwable $e) {
+            // We can't afford to throw exceptions in an exception handler :)
+            // Here, if the Retry-After header doesn't have a numeric value
+            // or a date that can be handled by DateTimeImmutable, we just
+            // throw it away, sorry not sorry ¯\_(ツ)_/¯
+            return null;
+        }
     }
 }