+++ src/Firebase/Messaging/ApiClient.php
@@ -27,10 +27,10 @@ class ApiClient implements ClientInterface
     /**
      * @internal
      */
-    public function __construct(ClientInterface $client)
+    public function __construct(ClientInterface $client, MessagingApiExceptionConverter $errorHandler)
     {
         $this->client = $client;
-        $this->errorHandler = new MessagingApiExceptionConverter();
+        $this->errorHandler = $errorHandler;
     }
 
     /**